+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.3.0 - October 12th 2020
 * Remove `aws_account_id` input variable
+* Update minimal example to create required resources
 
 ## v0.2.2 - July 6th 2020
 * Adds ListObjects to the s3 policy